Get script to run on icon click
A JS script now runs when we click the icon, or when we click the shortcut I've assigned via Chrome. In the long run, we need to set that shortcut ourselves, but it's a start.

The other long-term change is that right now we load jquery every time we want to run the script, and I suspect once the basic functionality is down I will try not to use jquery at all.