Finish popup optimization
Switched to generating HTML with strings and then using jquery only once.

The other main thing that bugs me is that I would prefer to load jquery on page load and trigger the rest of the extension later, so as to not have to load jquery every time. But couldn't seem to find an easy way to do that with content_scripts. May come back to it later.